:sparkles: feat: users can favorite pokemon
- Extract "favorite button" code and replace with new component in
PokemonListItems
- Use AppContext to store favorites across fetch calls when the
characters change
- With the divideIndexNumberByTwelves, we can get the appropriate index
number for the pokemonLists arrays (that are only a length of 12) to
save the pokemon to the favorite context array.
- The favorite heart button receives props from PokemonListItems
component, and toggles on and off when clicked. Favoriting a pokemon
stores them in a context array called yourFavoritedPokemon even when the
characters change due to a new API fetch call.

TODO:
- Characters cannot be removed from the array properly yet
- I can't quite get favorites to display properly after you navigate
away from the page they were loaded on initially. They are stored, but
visually, they don't display. You can also add multiples of the same
character (because they initially get "removed" when making new fetch
calls)